Split up sentinel classes
- coredis.sentinel.Sentinel -> coredis.client._sentinel.Sentinel
- coredis.sentinel.SentinelManagedConnection -> coredis.connection._sentinel.SentinelManagedConnection
- coredis.sentinel.SentinelConnectionPool -> coredis.pool._sentinel.SentinelConnectionPool
