+++ coredis/__init__.py
@@ -9,17 +9,22 @@ cluster & sentinel.
 from __future__ import annotations
 
 from coredis._version import __version__
-from coredis.client import Redis, RedisCluster
+from coredis.client import Redis, RedisCluster, Sentinel
 from coredis.config import Config
 from coredis.connection import (
     BaseConnection,
     ClusterConnection,
     Connection,
+    SentinelManagedConnection,
     TCPConnection,
     UnixDomainSocketConnection,
 )
-from coredis.pool import BaseConnectionPool, ClusterConnectionPool, ConnectionPool
-from coredis.sentinel import Sentinel
+from coredis.pool import (
+    BaseConnectionPool,
+    ClusterConnectionPool,
+    ConnectionPool,
+    SentinelConnectionPool,
+)
 from coredis.tokens import PureToken
 
 __all__ = [
@@ -36,5 +41,7 @@ __all__ = [
     "ClusterConnectionPool",
     "PureToken",
     "Sentinel",
+    "SentinelManagedConnection",
+    "SentinelConnectionPool",
     "__version__",
 ]
