+++ coredis/client/__init__.py
@@ -2,5 +2,6 @@ from __future__ import annotations
 
 from .basic import Client, Redis
 from .cluster import RedisCluster
+from .sentinel import Sentinel
 
-__all__ = ["Client", "Redis", "RedisCluster"]
+__all__ = ["Client", "Redis", "RedisCluster", "Sentinel"]
