+++ coredis/connection/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from ._base import BaseConnection, BaseConnectionParams, Location
 from ._cluster import ClusterConnection
+from ._sentinel import SentinelManagedConnection
 from ._tcp import TCPConnection, TCPLocation
 from ._uds import UnixDomainSocketConnection, UnixDomainSocketLocation
 
@@ -14,6 +15,7 @@ __all__ = [
     "TCPConnection",
     "Location",
     "TCPLocation",
+    "SentinelManagedConnection",
     "UnixDomainSocketLocation",
     "UnixDomainSocketConnection",
     "ClusterConnection",
