+++ coredis/pool/__init__.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from ._base import BaseConnectionPool, BaseConnectionPoolParams
 from ._basic import ConnectionPool, ConnectionPoolParams
 from ._cluster import ClusterConnectionPool, ClusterConnectionPoolParams
+from ._sentinel import SentinelConnectionPool
 
 __all__ = [
     "BaseConnectionPool",
@@ -11,4 +12,5 @@ __all__ = [
     "ConnectionPoolParams",
     "ClusterConnectionPool",
     "ClusterConnectionPoolParams",
+    "SentinelConnectionPool",
 ]
